MOVEONLY: Move backupwallet and restorewallet to rpc/backup.cpp
